pokemon_type_gradio_efficientNet/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
gradio>=4.0.0
pillow>=9.0.0
numpy>=1.24.0
pandas>=2.0.0